toolbar-actions specs - move to a named subdir
for consistency with other specs
