+++ app/javascript/spec/toolbar-actions/toolbar-actions-generic.test.js
@@ -1,6 +1,6 @@
 import 'angular-mocks';
-import getGridChecks from '../packs/toolbar-actions-common';
-import { onCustomAction } from '../toolbar-actions/custom-action';
+import getGridChecks from '../../packs/toolbar-actions-common';
+import { onCustomAction } from '../../toolbar-actions/custom-action';
 
 describe('Toolbar actions', () => {
   beforeEach(() => {
